Added client-side functions for getActivity, getFavorites, getPreferences
So the goal here was to display the user's activity on their activity feed. Unforuntately, in ClientTools.js, the getFavorites and getActivity functions are not working properly.

Assuming everything worked fine, I wanted to go into the UserProfilePage.js, then insert into the html and add the user's activity by adding the line {this.state.userActivity.activity}, which should display the user's activity according to the database.

Secondly, I also wanted to add the user's favorite recipes, so assuming the client-side function worked properly, I would go into the html and insert the line {this.state.userFavorites.favorites} to display the user's favorite recipes.

The respective functions are commented out, otherwise there is an error.